% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc}
\alias{sc}
\title{Poke or initiate Databricks connection}
\usage{
sc(hide_output = TRUE)
}
\arguments{
\item{hide_output}{logical. Hide the output of spark_connect(). Default TRUE}
}
\value{
The connection object, activating the connection
}
\description{
Should check if connection object already exists in the \code{.conns}
environment, and if so use it. If no connection found, it should initiate
the connection, and bind it to the \code{.conns} environment.
}
