% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.R
\name{sc_disconnect}
\alias{sc_disconnect}
\title{Disconnect Databricks connection}
\usage{
sc_disconnect()
}
\value{
TRUE if successful
}
\description{
If the connection object exists in the \code{.conns} environment, disconnect it.
}
